Substitute panic(...) with printk(...) in dbi_unregister_uretprobe
Sometimes the stack cannot be recovered after userspace instrumentation
because the corresponding function does not save LR value. In this case
it is better to print a warning instead of calling panic (the worst that
can happen is the instrumented application crashing with segmentation
fault). Currently the function _start is known to cause such an error.
But actually it never returns so it is save to ignore this issue and
just recycle the coresponding retprobe instance.