Added functionality for trap (undef. instr) overhead calculation (ARM)
When compiled with support of this functionality swap kprobe module will send
SIGUSR1 signal to userspace application after triggering 100000 traps
(undef.instructions). After that it can calculate the overhead of one trap
(transition from user to kernel and from kernel to user) as a difference of
time when catching signal SIGUSR1 in signal handler and time before first
undef.instruction which is divided on amount of traps (100000).