[FIX] stop message handling in us_manager
stop_machine is now used only to acquire references to all the
instrumented tasks. In such way we assure that they are not going
to be dropped during SWAP cleanup code execution.

Also a special callback STOP_CB_TD was added for task_data since it
may need to cleanup in stop_machine.

Change-Id: I4ea67092e1f0a006c908392e41ae511b9fff5b6d
Signed-off-by: Vyacheslav Cherkashin <v.cherkashin@samsung.com>